Move CatalogOperationsTest to `:openhouse-spark-itest` module  (#112)
Move CatalogOperationsTest to `:openhouse-spark-itest` module  (#112)
