+++ integrations/spark/openhouse-spark-itest/build.gradle
@@ -12,6 +12,8 @@ dependencies {
     exclude group: "io.netty"
   }
 
+  testImplementation(project(':tables-test-fixtures_2.12'))
+  testImplementation 'org.junit.platform:junit-platform-runner:1.8.2'
   // Required to test /tables mockserver
   testImplementation "com.squareup.okhttp3:okhttp:4.0.1"
   testImplementation "com.squareup.okhttp3:mockwebserver:4.0.1"
@@ -41,10 +43,18 @@ task statementTest(type: Test) {
   }
 }
 
+task catalogTest(type: Test) {
+  filter {
+    includeTestsMatching 'com.linkedin.openhouse.spark.catalogtest.*'
+  }
+}
+
 test {
   filter {
     excludeTestsMatching 'com.linkedin.openhouse.spark.statementtest.*'
+    excludeTestsMatching 'com.linkedin.openhouse.spark.catalogtest.*'
   }
 }
 
 test.dependsOn statementTest
+statementTest.dependsOn catalogTest
